google api implemented

+++ db/migrate/20170508175758_create_locations.rb
@@ -3,8 +3,8 @@ class CreateLocations < ActiveRecord::Migration
     create_table :locations do |t|
       t.string :name
       t.string :address
-      t.decimal :latitude
-      t.decimal :longitude
+      t.float :latitude
+      t.float :longitude
       t.references :pick_up, index: true, foreign_key: true
 
       t.timestamps null: false
